Add ability to delete proposal invitations
Some programs have the need to delete proposal invitations on a regular
basis. This pull request adds a view that allows the PI of a proposal to
delete invitations so that the invited user will no longer be added to
the proposal when signing up for an account (if they ever sign up for an
account).

Delivers https://www.pivotaltracker.com/story/show/161568643